Implement most of required protocols to make decoding responses to generic implementation.
There's follow types presented:
- enum `APIRequest` - main type that user should be working with.
- `APIResponse<T>` - type that returns network request method on successfully API call, where generic `T` type is the any given value of `result`.
- protocol `APIResponseType` - protocol that must conforms any possible `result` type.
- protocol `LiteralInitableFromString` - protocol that conforms any `Numeric` literals that `result` property hols as `String`.
- protocol `IntegerInitableWithRadix` - utility protocol that conforms any `Integer` literals that have follow initializee: `init?(from hexString: String)`
