--- Sources/web3swift/Web3/Web3+APIMethod.swift
@@ -1,8 +1,0 @@
-//
-//  Web3+APIMethod.swift
-//  Web3swift
-//
-//  Created by Yaroslav on 24.05.2022.
-//
-
-import Foundation
